+++ src/js/renderers/vimeo.js
@@ -305,6 +305,10 @@ const vimeoIframeRenderer = {
 				}
 			}
 
+			if (mediaElement.originalNode.muted) {
+				vimeoPlayer.setVolume(0);
+			}
+
 			const vimeoIframe = document.getElementById(vimeo.id);
 			let events;