+++ src/js/renderers/youtube.js
@@ -408,9 +408,6 @@ const YouTubeIframeRenderer = {
 					showinfo: 0,
 					modestbranding: 0,
 					html5: 1,
-					playsinline: 0,
-					start: 0,
-					end: 0,
 					iv_load_policy: 3
 				}, youtube.options.youtube),
 				origin: window.location.host,
@@ -443,7 +440,7 @@ const YouTubeIframeRenderer = {
 						youTubeIframe = youTubeApi.getIframe();
 
 						// Check for `muted` attribute to start video without sound
-						if (mediaElement.originalNode.getAttribute('muted')) {
+						if (mediaElement.originalNode.muted) {
 							youTubeApi.mute();
 						}
 
@@ -521,8 +518,9 @@ const YouTubeIframeRenderer = {
 			}
 		;
 
-		// The following will prevent that in mobile devices, YouTube is displayed in fullscreen when using audio
-		if (isAudio) {
+		// The following will prevent that, in mobile devices, YouTube is displayed in fullscreen when using audio
+		// of if the `playsinline` attribute is not set
+		if (isAudio || (!isAudio && mediaElement.originalNode.hasAttribute('playsinline'))) {
 			youtubeSettings.playerVars.playsinline = 1;
 		}